requirements.txt
Flask
pillow
pymongo
python-dotenv
opencv-python
numpy
requests

